refactor(dashboard): split memory.ts god file into 3 modules
876 lines → 3 focused modules:
- memory-state.ts (317): signals, data ops, vote/delete/comment handlers
- memory-post-detail.ts (262): PostDetail, CommentThread, CommentForm
- memory.ts (404): Memory main, PostCard, SortBar, NewPostForm

Dependency: memory → state + post-detail, post-detail → state. No cycles.
